Fixed pr-review comments.
